CASMPET-7238: /etc/target/saveconfig.json exists as a dir instead of file
The iscsi configuration is saved in /etc/target/saveconfig.json. But
on some systems this exists as directory and actual file is inside this
directory as /etc/target/saveconfig.json/saveconfig.json.temp Hence this
case needs to handled in SBPS Marshal agent, else SBPS Marshal agent
will fail. This code fix is addressing this issue.